+++ ocaml-lsp-server/src/ocaml_lsp_server.ml
@@ -166,6 +166,7 @@ let initialize_info (client_capabilities : ClientCapabilities.t) : InitializeRes
       ~declarationProvider:(`Bool true)
       ~definitionProvider:(`Bool true)
       ~typeDefinitionProvider:(`Bool true)
+      ~implementationProvider:(`Bool true)
       ~completionProvider
       ~signatureHelpProvider
       ~codeActionProvider
@@ -764,6 +765,8 @@ let on_request
     later (fun state () -> Definition_query.run `Definition state uri position) ()
   | TextDocumentTypeDefinition { textDocument = { uri }; position; _ } ->
     later (fun state () -> Definition_query.run `Type_definition state uri position) ()
+  | TextDocumentImplementation { textDocument = { uri }; position; _ } ->
+    later (fun state () -> Implementation_query.run state uri position) ()
   | TextDocumentCompletion params -> later (fun _ () -> Compl.complete state params) ()
   | TextDocumentPrepareRename req ->
     later
@@ -814,7 +817,6 @@ let on_request
          let+ result = selection_range state req in
          Some result)
       req
-  | TextDocumentImplementation _ -> Server.not_supported ()
   | SemanticTokensFull p -> later Semantic_highlighting.on_request_full p
   | SemanticTokensDelta p -> later Semantic_highlighting.on_request_full_delta p
   | TextDocumentMoniker _ -> Server.not_supported ()